refactor: rename SelectPeopleUI to SelectPeopleAndGroupsUI
