+++ src/components/shared/SelectPeopleAndGroups/index.tsx
@@ -1,3 +1,3 @@
-export * from './SelectPeopleUI'
+export * from './SelectPeopleAndGroupsUI'
 export * from './PersonOrGroupInList'
 export * from './PersonOrGroupInChip'
