+++ src/stories/index.tsx
@@ -3,7 +3,7 @@ import './OptionsPage'
 import './Injections'
 import './shared'
 import { definedSocialNetworkUIs, defineSocialNetworkUI, activateSocialNetworkUI } from '../social-network/ui'
-import { demoPeople, demoGroup } from './demoPeople'
+import { demoPeople, demoGroup } from './demoPeopleOrGroups'
 import { ValueRef } from '@holoflows/kit/es'
 import { PersonIdentifier } from '../database/type'
 import { emptyDefinition } from '../social-network/defaults/emptyDefinition'
